moved filter map to util
