+++ app/src/main/java/com/example/unlibrary/unlibrary/UnlibraryViewModel.java
@@ -21,7 +21,7 @@ import androidx.navigation.NavDirections;
 import androidx.navigation.Navigation;
 
 import com.example.unlibrary.book_list.BooksSource;
-import com.example.unlibrary.library.FilterMap;
+import com.example.unlibrary.util.FilterMap;
 import com.example.unlibrary.models.Book;
 import com.example.unlibrary.models.Request;
 import com.example.unlibrary.util.BarcodeScanner;
